chore(webpack): ignore ts2321 error
new ramda types causes an error with curry:
Type error: Excessive stack depth comparing types 'GapsOf<?, T2>' and
'GapsOf<?, T2>'.  TS2321
this error could be thrown the a specific call of curry but isnt found
yet. Temporarily ignore this error to be able to use ts3.7-beta.
The problem was traced down until the Curry type definition in the
ts-toolbelt package. Disabling the Curry definition there removed the
error but with the downside of having no Curry types at all.